Версия 1.2.5 Основное - проведена ревизия и значительные изменения кода в информационном обмене. Изменена структура БД. Требуется обновить версию
git-svn-id: svn+ssh://svn.code.sf.net/p/dynamicdocs/code@691 05742d63-fc59-4ccd-b3f0-d7e2a2a8a88d

+++ ddocclient/app/dyndoc_synch/dds_http_connector/src/httpwindow.h
@@ -83,7 +83,7 @@ private:
 
     QList<JKKSPMessWithAddr *> messageList;
 
-    QMap<int, QPair<int, int> > httpMessages;
+    QMap<int, QPair<qint64, qint64> > httpMessages;
 
     //QList< QPair<int, int> > http_message;
     //QList< QPair<int, int> > message_type;
