+++ ddocclient/kksinteractor/JKKSCategory.h
@@ -29,23 +29,23 @@ class JKKSLoader;
 class _I_EXPORT JKKSCategory : public JKKSUID
 {
     public:
-        JKKSCategory (int idCat = -1, 
+        JKKSCategory (qint64 idCat = -1, 
                       const QString& cName=QString(), 
                       const QString& cCode=QString(), 
                       const QString& cDesc=QString(), 
                       int type=-1, 
-                      int id_child=-1, 
+                      qint64 id_child=-1, 
                       bool isCMain=false,
                       const QString & uid = QString(),
-                      int id_io_state=1,
+                      qint64 id_io_state=1,
                       bool is_global = false);
         //JKKSCategory (int idCat, const QString& cCode, const QString & uid = QString());
         JKKSCategory (const JKKSCategory& cat);
         ~JKKSCategory (void);
         JKKSCategory& operator=(const JKKSCategory& that);
 
-        int id (void) const;
-        void setId (int idCat);
+        qint64 id (void) const;
+        void setId (qint64 idCat);
 
         QString getName (void) const;
         void setName (const QString& cname);
@@ -59,11 +59,11 @@ class _I_EXPORT JKKSCategory : public JKKSUID
         int getType (void) const;
         void setType (int type);
 
-        int getState (void) const;
-        void setState (int type);
+        qint64 getState (void) const;
+        void setState (qint64 type);
 
-        int getIDChild (void) const;
-        void setIDChild (int idc);
+        qint64 getIDChild (void) const;
+        void setIDChild (qint64 idc);
 
         bool isMain (void) const;
         void setMain (bool val);
@@ -71,13 +71,13 @@ class _I_EXPORT JKKSCategory : public JKKSUID
         bool isGlobal (void) const;
         void setGlobal (bool val);
 
-        const QMap<int, JKKSCategoryAttr>& attributes (void) const;
-        QMap<int, JKKSCategoryAttr>& attributes (void);
-        void setAttributes (const QMap<int, JKKSCategoryAttr>& attrs);
+        const QMap<qint64, JKKSCategoryAttr>& attributes (void) const;
+        QMap<qint64, JKKSCategoryAttr>& attributes (void);
+        void setAttributes (const QMap<qint64, JKKSCategoryAttr>& attrs);
 
-        const QMap<int, JKKSRubric>& rubrics (void) const;
-        QMap<int, JKKSRubric>& rubrics (void);
-        void setRubrics (const QMap<int, JKKSRubric>& rubrs);
+        const QMap<qint64, JKKSRubric>& rubrics (void) const;
+        QMap<qint64, JKKSRubric>& rubrics (void);
+        void setRubrics (const QMap<qint64, JKKSRubric>& rubrs);
 
     private:
         //
@@ -91,18 +91,18 @@ class _I_EXPORT JKKSCategory : public JKKSUID
         //
         // Variables
         //
-        int id_io_category;
+        qint64 id_io_category;
         QString m_catName;
         QString m_catCode;
         QString m_catDesc;
         int m_idType;
-        int m_idState;
-        int m_idChild;
+        qint64 m_idState;
+        qint64 m_idChild;
         bool m_isMain;
         bool m_isGlobal;
 
-        QMap<int, JKKSCategoryAttr> m_attributes;
-        QMap<int, JKKSRubric> m_rubrics;
+        QMap<qint64, JKKSCategoryAttr> m_attributes;
+        QMap<qint64, JKKSRubric> m_rubrics;
 };
 
 QDataStream& operator<< (QDataStream& out, const JKKSCategory& cat);
