+++ ddocclient/kksinteractor/JKKSGlobalRubric.cpp
@@ -1,15 +1,15 @@
 #include <QDataStream>
 #include "JKKSGlobalRubric.h"
 
-JKKSGlobalRubric :: JKKSGlobalRubric (int id,
+JKKSGlobalRubric :: JKKSGlobalRubric (qint64 id,
                                       const QString & uid,
-                                      int idp,
-                                      int idc,
+                                      qint64 idp,
+                                      qint64 idc,
                                       const QString& name,
                                       const QString& code,
                                       const QString& desc)
     : JKKSRubric (id, uid, idp, idc, name, code, desc),
-    category (QMap<int, JKKSCategory>()),
+    category (QMap<qint64, JKKSCategory>()),
     searchTemplate (JKKSSearchTemplate())
 {
 }
@@ -33,17 +33,17 @@ JKKSGlobalRubric& JKKSGlobalRubric :: operator= (const JKKSGlobalRubric& R)
     return *this;
 }
 
-const QMap<int, JKKSCategory>& JKKSGlobalRubric :: getCategory (void) const
+const QMap<qint64, JKKSCategory>& JKKSGlobalRubric :: getCategory (void) const
 {
     return category;
 }
 
-QMap<int, JKKSCategory>& JKKSGlobalRubric :: getCategory (void)
+QMap<qint64, JKKSCategory>& JKKSGlobalRubric :: getCategory (void)
 {
     return category;
 }
 
-void JKKSGlobalRubric :: setCategory (const QMap<int, JKKSCategory>& c)
+void JKKSGlobalRubric :: setCategory (const QMap<qint64, JKKSCategory>& c)
 {
     category = c;
 }
