+++ ddocclient/kksinteractor/JKKSIOTable.h
@@ -21,7 +21,7 @@ class JKKSCategory;
 class _I_EXPORT JKKSIOTable : public JKKSUID
 {
 public:
-    JKKSIOTable (int idIOTable = -1,
+    JKKSIOTable (qint64 idIOTable = -1,
                  const QString & table_name = QString(),
 //                 const QMap<int, JKKSCategory>& catMap,// = QMap<int, JKKSCategory>(),
                  const QString & title = QString(),
@@ -31,14 +31,14 @@ public:
 
     ~JKKSIOTable (void);
 
-    int getIDTable (void) const;
-    void setIDTable (int _id);
+    qint64 getIDTable (void) const;
+    void setIDTable (qint64 _id);
 
     QString getName (void) const;
     void setName (QString tName);
 
-    const QMap<int, JKKSCategory>& getCategory (void) const;
-    void setCategory (const QMap<int, JKKSCategory>& catMap);
+    const QMap<qint64, JKKSCategory>& getCategory (void) const;
+    void setCategory (const QMap<qint64, JKKSCategory>& catMap);
 
     QString getTitle (void) const;
     void setTitle (QString tTitle);
@@ -54,9 +54,9 @@ private:
     //
     // Variables
     //
-    int id;
+    qint64 id;
     QString tableName;
-    QMap<int, JKKSCategory> category;
+    QMap<qint64, JKKSCategory> category;
     QString Title;
 };
